requirements.txt
absl-py==0.9.0
amqp==2.5.2
aniso8601==8.0.0
appdirs==1.4.3
apturl==0.5.2
asn1crypto==0.24.0
astor==0.8.1
attrs==19.3.0
backcall==0.1.0
bcrypt==3.1.7
beautifulsoup4==4.9.1
billiard==3.6.3.0
bleach==3.1.5
Brlapi==0.6.6
cached-property==1.5.1
celery==4.4.2
certifi==2020.6.20
cffi==1.12.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
command-not-found==0.3
configparser==5.0.0
contextlib2==0.6.0.post1
crayons==0.4.0
cryptography==2.7
cupshelpers==1.0
cycler==0.10.0
Cython==0.29.17
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
Deprecated==1.2.7
distro-info===0.18ubuntu0.18.04.1
Django==2.0.5
docker==3.7.3
docker-compose==1.24.1
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.3
Events==0.3
Flask==1.1.2
Flask-RESTful==0.3.8
gast==0.3.3
glog==0.3.1
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
httplib2==0.9.2
idna==2.10
imageio==2.9.0
IMDbPY==6.8
importlib-metadata==1.6.0
ipykernel==5.2.1
ipython==7.14.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
kazam==1.4.5
keep==2.7
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.2.0
kombu==4.6.8
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
lxml==4.5.2
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.6
netifaces==0.10.4
notebook==6.0.3
numpy==1.19.1
oauth==1.0.1
olefile==0.45.1
opencv-python==4.1.1.26
packaging==20.3
pandas==0.25.3
pandocfilters==1.4.2
paramiko==2.6.0
parso==0.7.0
pascal-voc-writer==0.1.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.12.0
ptyprocess==0.6.0
pycairo==1.16.2
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
PyGithub==1.45
Pygments==2.6.1
pygobject==3.26.1
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.7
PyQt5==5.14.0
PyQt5-sip==12.7.0
pyRFC3339==1.0
pyrsistent==0.16.0
python-apt==1.6.5+ubuntu0.5
python-dateutil==2.8.1
python-debian==0.1.32
python-gflags==3.1.2
pytz==2020.1
pyxdg==0.25
PyYAML==5.3.1
pyzmq==19.0.1
qtconsole==4.7.4
QtPy==1.9.0
reportlab==3.4.0
requests==2.24.0
requests-unixsocket==0.1.5
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
scipy==1.4.1
SecretStorage==2.3.1
selenium==3.141.0
Send2Trash==1.5.0
simplejson==3.13.2
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.19
ssh-import-id==5.7
system-service==0.3
systemd-python==234
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.3
terminaltables==3.1.0
testpath==0.4.4
texttable==0.9.1
tf-slim==1.1.0
tornado==6.0.4
traitlets==4.3.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.10
usb-creator==0.3.3
vine==1.3.0
virtualenv==16.7.7
wadllib==1.3.2
wcwidth==0.1.9
webdriver-manager==3.2.2
webencodings==0.5.1
websocket-client==0.56.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xkit==0.0.0
zipp==3.1.0
zope.interface==4.3.2
